refactor(imports): change circular import references
