+++ src/AvatarStack.tsx
@@ -2,7 +2,7 @@ import classnames from 'classnames'
 import React from 'react'
 import styled from 'styled-components'
 import {get} from './constants'
-import {Box} from '.'
+import Box from './Box'
 import sx, {SxProp} from './sx'
 
 type StyledAvatarStackWrapperProps = {
