+++ src/FormControl/FormControl.tsx
@@ -1,5 +1,13 @@
 import React, {useContext} from 'react'
-import {Autocomplete, Box, Checkbox, Radio, Select, Textarea, TextInput, TextInputWithTokens, useSSRSafeId} from '..'
+import Autocomplete from '../Autocomplete'
+import Box from '../Box'
+import Checkbox from '../Checkbox'
+import Radio from '../Radio'
+import Select from '../Select'
+import Textarea from '../Textarea'
+import TextInput from '../TextInput'
+import TextInputWithTokens from '../TextInputWithTokens'
+import {useSSRSafeId} from '../utils/ssr'
 import FormControlCaption from './_FormControlCaption'
 import FormControlLabel from './_FormControlLabel'
 import FormControlValidation from './_FormControlValidation'
