+++ src/SegmentedControl/SegmentedControlButton.tsx
@@ -1,7 +1,7 @@
 import React, {HTMLAttributes} from 'react'
 import {IconProps} from '@primer/octicons-react'
 import styled from 'styled-components'
-import {Box} from '..'
+import Box from '../Box'
 import sx, {merge, SxProp} from '../sx'
 import {getSegmentedControlButtonStyles, getSegmentedControlListItemStyles} from './getSegmentedControlStyles'
