+++ src/Timeline.tsx
@@ -1,7 +1,7 @@
 import classnames from 'classnames'
 import React from 'react'
 import styled, {css} from 'styled-components'
-import {Box} from '.'
+import Box from './Box'
 import {get} from './constants'
 import sx, {SxProp} from './sx'
 import {ComponentProps} from './utils/types'
