+++ src/ToggleSwitch.tsx
@@ -1,7 +1,9 @@
 import React, {MouseEventHandler, useCallback, useEffect} from 'react'
 import styled, {css} from 'styled-components'
 import {variant} from 'styled-system'
-import {Box, Spinner, Text} from '.'
+import Box from './Box'
+import Spinner from './Spinner'
+import Text from './Text'
 import {get} from './constants'
 import {useProvidedStateOrCreate} from './hooks'
 import sx, {BetterSystemStyleObject, SxProp} from './sx'
