+++ src/_InputLabel.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import {Box} from '.'
+import Box from './Box'
 import {SxProp} from './sx'
 import VisuallyHidden from './_VisuallyHidden'
