+++ src/_InputValidation.tsx
@@ -1,6 +1,7 @@
 import {AlertFillIcon, CheckCircleFillIcon, IconProps} from '@primer/octicons-react'
 import React from 'react'
-import {Box, Text} from '.'
+import Box from './Box'
+import Text from './Text'
 import {SxProp} from './sx'
 import {FormValidationStatus} from './utils/types/FormValidationStatus'
