+++ src/_TextInputInnerAction.tsx
@@ -1,7 +1,8 @@
 import React, {forwardRef} from 'react'
 import {IconProps} from '@primer/octicons-react'
-import {Box, Button, IconButton, Tooltip} from '.'
-import {ButtonProps} from './Button'
+import Box from './Box'
+import {Button, IconButton, ButtonProps} from './Button'
+import Tooltip from './Tooltip'
 import {BetterSystemStyleObject, merge, SxProp} from './sx'
 
 type TextInputActionProps = Omit<React.HTMLProps<HTMLButtonElement>, 'aria-label' | 'size'> & {
